+++ src/index.js
@@ -1,4 +1,5 @@
 import React from "react";
+import "./App.css";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App";
